Increase delay before copying asoundrc
At least 1 of our pis is showing that 15 seconds is too short. I tried 20 seconds and that fixed it.

A better solution would be to use pulseaudio instead of ALSA to play and record. I think it is worth investigation as this is the behaviour that the OS is expecting of us and I imagine this 20 second increase will not fix the problem forever